requirements.txt
# ===== inbound =========
boto3==1.34.11
psycopg2-binary==2.9.9
aws_logging_handlers==2.0.5
watchtower==3.0.1
sqlalchemy==2.0.30
requests==2.31.0
botocore==1.34.11
s3transfer==0.10.0
urllib3==2.0.7
aiohttp

# === performace / core ===
objgraph

# ==== web ====
uvicorn==0.32.1
fastapi==0.111.0
aiofiles==23.2.1
fastapi-cache2==0.2.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# =====Oauth (github) ====
authlib
python-dotenv

# === rate limiteing + caching
#slowapi
fastapi-limiter
redis
redis[asyncio]